requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.10.4
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.5
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
distro==1.9.0
dnspython==1.16.0
et_xmlfile==2.0.0
eventlet==0.30.2
exceptiongroup==1.3.0
fastapi==0.116.1
Flask==2.3.3
flask-cors==6.0.1
Flask-SocketIO==5.3.6
frozenlist==1.7.0
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.0
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
lz4==4.4.4
MarkupSafe==3.0.3
motor==3.7.1
multidict==6.6.4
numpy==2.0.2
oauth2client==4.1.3
oauthlib==3.3.1
openai==1.108.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.2
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
playwright==1.55.0
praw==7.8.1
prawcore==2.4.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
protoc-gen-openapiv2==0.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pyee==13.0.0
pymongo==4.15.0
pyparsing==3.2.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.3
python-socketio==5.14.0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
selenium==4.15.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
starlette==0.47.3
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
update-checker==0.18.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
webdriver-manager==4.0.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==2.3.7
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
